requirements.txt
pillow>=10.0.0
exif>=1.6.0
pyexiftool>=0.5.6
pymediainfo>=6.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
geopy>=2.0.0
python-dateutil>=2.8.0
questionary>=2.0.0
colorama>=0.4.0
pathvalidate>=3.0.0
numpy>=1.24.0
pyinstaller>=6.0.0
matplotlib>=3.7.0
imagehash>=4.3.0
send2trash>=1.8.0